Tidy some tests up in order to update hypothesis (#103)
* Update tests to work with latest Hypothesis

Simplify tests using the builtin operator module
to remove some lambdas. Refactor all operator
test cases so that the @given decorator is applied
only to test methods - this removes the do_test*
methods and lets Hypothesis work with the out of
the box configuration.

* Remove Hypothesis upper bound pins, improve test reporting

Adding -v to test reporting and setting the PYTHONFAULTHANDLER
environment variable in tox. This will enable better reporting
of failures, even when the failure is a segfault.