Add type stubs (& minimal validation)
This adds type stubs to the published package, so that consumers
can benefit from type checking despite the Cython implementation
of this package. The stubs are manually constructed as there does
not appear to be a mature solution which produces good stubs from
Cython code. (It is actually unclear if such a solution is
possible without annotation of the Cython source)

Validation of the stubs is achieved by annotating the tests and
relying on the existing test coverage to explore all relevant
members & operations. A new tox environment runs mypy against the
tests.

Fixes https://github.com/Ezibenroc/PyRoaringBitMap/issues/105