Hoist helper which isn't a method
Mypy complains that this doesn't have a self/cls argument, however
adding @staticmethod causes test failures on Python < 3.10 as the
resulting object isn't callable (presumably because the class hasn't
finished initiating when we call the method). Hoisting this to a
function avoids the issue entirely, at the cost of this helper
being slightly further from its usages.