Improve type handling in rich comparisons
Previously an error was raised when doing equality comparisons
between BitMaps and other instance types, which is unexpected in
Python. Move over to implementing each of the rich comparison
methods directly so that we can more match the usual pattern of
allowed (and rejected) comparisons.

Also explicitly check for `None` when validating that two BitMaps
can be compared, so that the error message emitted in that case
is of the more expected type (i.e: `TypeError` rather than
`AttributeError`).